Split out export from ShowGlobals into ExportGlobals - showglobals only prints to the screen - exportglobals only saves to file
