+++ src/main/scala/org/broadinstitute/hail/driver/Command.scala
@@ -62,6 +62,7 @@ object ToplevelCommands {
   register(ExportPlink)
   register(ExportGEN)
   register(ExportGenotypes)
+  register(ExportGlobals)
   register(ExportSamples)
   register(ExportVariants)
   register(ExportVariantsCass)
@@ -103,7 +104,7 @@ object ToplevelCommands {
   register(SampleQC)
   register(SeqrServerCommand)
   register(PrintSchema)
-  register(ShowGlobalAnnotations)
+  register(ShowGlobals)
   register(VariantQC)
   register(VEP)
   register(Write)
