--- src/main/scala/org/broadinstitute/hail/driver/ShowGlobalAnnotations.scala
@@ -1,41 +1,0 @@
-package org.broadinstitute.hail.driver
-
-import org.broadinstitute.hail.utils._
-import org.broadinstitute.hail.expr._
-import org.json4s.jackson.JsonMethods._
-import org.kohsuke.args4j.{Option => Args4jOption}
-
-object ShowGlobalAnnotations extends Command {
-
-  class Options extends BaseOptions {
-    @Args4jOption(name = "-o", aliases = Array("--output"), usage = "Output file")
-    var output: String = _
-  }
-
-  def newOptions = new Options
-
-  def name = "showglobals"
-
-  def description = "Show the signatures for all annotations currently stored in the dataset"
-
-  def supportsMultiallelic = true
-
-  def requiresVDS = true
-
-  def run(state: State, options: Options): State = {
-    val vds = state.vds
-
-    val json = vds.globalSignature.toJSON(vds.globalAnnotation)
-    val result = if (vds.globalSignature == TStruct.empty)
-      "Global annotations: `global' = Empty"
-    else
-      "Global annotations: `global' = " + pretty(render(json))
-
-    options.output match {
-      case null => println(result)
-      case path => state.hadoopConf.writeTextFile(path) { out => out.write(result) }
-    }
-
-    state
-  }
-}
